Move MockBuilder into own file
