+++ test/mocks/mock-builder.js
@@ -1,0 +1,20 @@
+'use strict';
+
+function MockBuilder(envelope, message) {
+    this.envelope = envelope;
+    this.message = message;
+}
+
+MockBuilder.prototype.getEnvelope = function() {
+    return this.envelope;
+};
+
+MockBuilder.prototype.createReadStream = function() {
+    return this.message;
+};
+
+MockBuilder.prototype.getHeader = function() {
+    return 'teretere';
+};
+
+module.exports = MockBuilder;
